Compare reconnect epoch seconds as longs in ChangedFieldsTest
A float holds epoch seconds from 2004 to 2038 only in steps of 128, so
two readings one second apart could parse 128 apart and fail the
10-second tolerance.

+++ src/test/org/epics/archiverappliance/engine/V4/ChangedFieldsTest.java
@@ -277,7 +277,7 @@ public class ChangedFieldsTest {
             for (var v : e.getValue().entrySet()) {
                 if (v.getKey().equals("cnxregainedepsecs")) {
                     Assertions.assertTrue(
-                            Math.abs(Float.parseFloat(v.getValue()) - Float.parseFloat(actualMap.get(v.getKey())))
+                            Math.abs(Long.parseLong(v.getValue()) - Long.parseLong(actualMap.get(v.getKey())))
                                     < 10);
                 } else {
                     Assertions.assertEquals(v.getValue(), actualMap.get(v.getKey()));
